gateway_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx # La biblioteca clave para hacer llamadas a otros servicios